+++ src/client/java/io/github/tr100000/codec2schema/ClientDelegateClientImpl.java
@@ -2,19 +2,15 @@ package io.github.tr100000.codec2schema;
 
 import org.jspecify.annotations.NullMarked;
 
-import java.util.List;
-
 @NullMarked
 public class ClientDelegateClientImpl extends ClientDelegate {
-    private static final List<String> ENTRYPOINT_KEYS = List.of("codec2schema:main", "codec2schema:client");
-
     @Override
     public void registerHandlers() {
-        Codec2Schema.registerHandlers(ENTRYPOINT_KEYS);
+        Codec2Schema.registerHandlers(PluginSide.MAIN, PluginSide.CLIENT);
     }
 
     @Override
     public void generateSchemas() {
-        Codec2Schema.generateSchemas(ENTRYPOINT_KEYS);
+        Codec2Schema.generateSchemas(PluginSide.MAIN, PluginSide.CLIENT);
     }
 }