+++ src/main/java/io/github/tr100000/codec2schema/impl/specific/WrappedTypedEntityDataCodec.java
@@ -10,6 +10,7 @@ import org.jspecify.annotations.NullMarked;
 @NullMarked
 public interface WrappedTypedEntityDataCodec<IdType> extends Codec<TypedEntityData<IdType>> {
     Codec<TypedEntityData<IdType>> original();
+    Codec<IdType> typeCodec();
 
     @Override
     default <T> DataResult<Pair<TypedEntityData<IdType>, T>> decode(DynamicOps<T> ops, T input) {