+++ src/main/java/io/github/tr100000/codec2schema/mixin/CodecMixin.java
@@ -11,8 +11,6 @@ import com.mojang.serialization.DynamicOps;
 import com.mojang.serialization.Encoder;
 import com.mojang.serialization.Lifecycle;
 import com.mojang.serialization.MapCodec;
-import com.mojang.serialization.MapDecoder;
-import com.mojang.serialization.MapEncoder;
 import com.mojang.serialization.MapLike;
 import com.mojang.serialization.RecordBuilder;
 import io.github.tr100000.codec2schema.api.NumberBound;
@@ -34,34 +32,33 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 @Mixin(Codec.class)
 public interface CodecMixin<A> {
     @WrapOperation(method = "xmap", at = @At(value = "INVOKE", target = "Lcom/mojang/serialization/Codec;of(Lcom/mojang/serialization/Encoder;Lcom/mojang/serialization/Decoder;Ljava/lang/String;)Lcom/mojang/serialization/Codec;"))
-    private <S> Codec<S> xmap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends S> toFunction) {
-        return wrapped(encoder, decoder, name, value -> DataResult.success(toFunction.apply(value)));
+    private <S> Codec<S> xmap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends S> to) {
+        return wrapped(encoder, decoder, name, value -> DataResult.success(to.apply(value)));
     }
 
     @WrapOperation(method = "comapFlatMap", at = @At(value = "INVOKE", target = "Lcom/mojang/serialization/Codec;of(Lcom/mojang/serialization/Encoder;Lcom/mojang/serialization/Decoder;Ljava/lang/String;)Lcom/mojang/serialization/Codec;"))
-    private <S> Codec<S> comapFlatMap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends DataResult<? extends S>> toFunction) {
-        return wrapped(encoder, decoder, name, toFunction);
+    private <S> Codec<S> comapFlatMap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends DataResult<? extends S>> to) {
+        return wrapped(encoder, decoder, name, to);
     }
 
     @WrapOperation(method = "flatComapMap", at = @At(value = "INVOKE", target = "Lcom/mojang/serialization/Codec;of(Lcom/mojang/serialization/Encoder;Lcom/mojang/serialization/Decoder;Ljava/lang/String;)Lcom/mojang/serialization/Codec;"))
-    private <S> Codec<S> flatComapMap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends S> toFunction) {
-        return wrapped(encoder, decoder, name, value -> DataResult.success(toFunction.apply(value)));
+    private <S> Codec<S> flatComapMap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends S> to) {
+        return wrapped(encoder, decoder, name, value -> DataResult.success(to.apply(value)));
     }
 
     @WrapOperation(method = "flatXmap", at = @At(value = "INVOKE", target = "Lcom/mojang/serialization/Codec;of(Lcom/mojang/serialization/Encoder;Lcom/mojang/serialization/Decoder;Ljava/lang/String;)Lcom/mojang/serialization/Codec;"))
-    private <S> Codec<S> flatXmap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends DataResult<? extends S>> toFunction) {
-        return wrapped(encoder, decoder, name, toFunction);
+    private <S> Codec<S> flatXmap(Encoder<S> encoder, Decoder<S> decoder, String name, Operation<Codec<A>> original, @Local(argsOnly = true, ordinal = 0) Function<? super A, ? extends DataResult<? extends S>> to) {
+        return wrapped(encoder, decoder, name, to);
     }
 
     @Unique
     @SuppressWarnings("unchecked")
-    private <S> Codec<S> wrapped(final Encoder<S> encoder, final Decoder<S> decoder, final String name, final Function<? super A, ? extends DataResult<? extends S>> toFunction) {
+    private <S> Codec<S> wrapped(final Encoder<S> encoder, final Decoder<S> decoder, final String name, final Function<? super A, ? extends DataResult<? extends S>> to) {
         Codec<A> thisCodec = (Codec<A>)this;
         Codec<S> fakeThisCodec = (Codec<S>)this;
         return switch (thisCodec) {
@@ -69,7 +66,7 @@ public interface CodecMixin<A> {
                 @Override
                 public List<ValueStringPair<S>> possibleValues() {
                     return codecWithValuePairs.possibleValues().stream()
-                            .map(pair -> pair.mapValue(toFunction))
+                            .map(pair -> pair.mapValue(to))
                             .filter(pair -> pair.value().hasResultOrPartial())
                             .map(pair -> pair.mapValue(DataResult::getPartialOrThrow))
                             .map(pair -> (ValueStringPair<S>)pair)
@@ -125,11 +122,12 @@ public interface CodecMixin<A> {
         };
     }
 
-    @WrapOperation(method = "fieldOf(Ljava/lang/String;)Lcom/mojang/serialization/MapCodec;", at = @At(value = "INVOKE", target = "Lcom/mojang/serialization/MapCodec;of(Lcom/mojang/serialization/MapEncoder;Lcom/mojang/serialization/MapDecoder;Ljava/util/function/Supplier;)Lcom/mojang/serialization/MapCodec;"))
+    @Inject(method = "fieldOf(Ljava/lang/String;)Lcom/mojang/serialization/MapCodec;", at = @At("RETURN"), cancellable = true)
     @SuppressWarnings("unchecked")
-    private MapCodec<A> fieldOf(MapEncoder<A> encoder, MapDecoder<A> decoder, Supplier<String> name, Operation<MapCodec<A>> original, @Local(argsOnly = true, ordinal = 0) final String fieldName) {
+    private void fieldOf(String name, CallbackInfoReturnable<MapCodec<A>> cir) {
         Codec<A> thisCodec = (Codec<A>)this;
-        return new WrappedFieldMapCodec<>() {
+        MapCodec<A> wrappedCodec = cir.getReturnValue();
+        cir.setReturnValue(new WrappedFieldMapCodec<>() {
             @Override
             public Codec<A> original() {
                 return thisCodec;
@@ -137,29 +135,29 @@ public interface CodecMixin<A> {
 
             @Override
             public String fieldName() {
-                return fieldName;
+                return name;
             }
 
             @Override
             public <T> Stream<T> keys(DynamicOps<T> ops) {
-                return Stream.concat(encoder.keys(ops), decoder.keys(ops));
+                return wrappedCodec.keys(ops);
             }
 
             @Override
-            public <T> DataResult<A> decode(final DynamicOps<T> ops, final MapLike<T> input) {
-                return decoder.decode(ops, input);
+            public <T> DataResult<A> decode(DynamicOps<T> ops, MapLike<T> input) {
+                return wrappedCodec.decode(ops, input);
             }
 
             @Override
-            public <T> RecordBuilder<T> encode(final A input, final DynamicOps<T> ops, final RecordBuilder<T> prefix) {
-                return encoder.encode(input, ops, prefix);
+            public <T> RecordBuilder<T> encode(A input, DynamicOps<T> ops, RecordBuilder<T> prefix) {
+                return wrappedCodec.encode(input, ops, prefix);
             }
 
             @Override
             public String toString() {
-                return name.get();
+                return wrappedCodec.toString();
             }
-        };
+        });
     }
 
     @Inject(method = "optionalFieldOf(Ljava/lang/String;Ljava/lang/Object;Z)Lcom/mojang/serialization/MapCodec;", at = @At("RETURN"), cancellable = true)